backend/requirements.txt
Django 
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==5.0.2
django-allauth==0.57.2